# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: JobQuest API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      summary: List jobs with filters
      parameters:
        - name: search
          in: query
          schema: { type: string }
        - name: location
          in: query
          schema: { type: string }
        - name: workMode
          in: query
          schema: { type: string, enum: [remote, hybrid, onsite] }
        - name: source
          in: query
          schema: { type: string }
        - name: experience
          in: query
          schema: { type: string }
        - name: salaryMin
          in: query
          schema: { type: integer }
        - name: postedWithin
          in: query
          schema: { type: string, enum: [1h, 6h, 12h, 24h, 3d, 7d] }
        - name: skills
          in: query
          schema: { type: string }
        - name: sortBy
          in: query
          schema: { type: string, enum: [latest, salary, company, matchScore] }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: List of jobs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobListResponse"
...
  /jobs/{id}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: Get a single job
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Job details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "404":
          description: Not found
  /jobs/{id}/similar:
    get:
      operationId: getSimilarJobs
      tags: [jobs]
      summary: Get similar jobs
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Similar jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Job"
  /jobs/{id}/save:
    post:
      operationId: saveJob
      tags: [jobs]
      summary: Save a job for later
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedJobResult"

    delete:
      operationId: unsaveJob
      tags: [jobs]
      summary: Remove job from saved
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Removed
...
  /resumes:
    get:
      operationId: listResumes
      tags: [resumes]
      summary: List user's resumes
      responses:
        "200":
          description: Resumes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Resume"

    post:
      operationId: createResume
      tags: [resumes]
      summary: Create/upload a resume record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResumeInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resume"
  /resumes/{id}:
    get:
      operationId: getResume
      tags: [resumes]
      summary: Get resume details
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Resume
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resume"

    patch:
      operationId: updateResume
      tags: [resumes]
      summary: Update resume
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResumeUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resume"

    delete:
      operationId: deleteResume
      tags: [resumes]
      summary: Delete resume
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ─── ATS ────────────────────────────────────────────────────────────────────
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List notifications
      parameters:
        - name: unreadOnly
          in: query
          schema: { type: boolean }
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
components:
  schemas:
...
    Job:
      type: object
      required: [id, title, company, location, source, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        company:
          type: string
        companyLogo:
          type: ["string", "null"]
        location:
          type: string
        workMode:
          type: string
          enum: [remote, hybrid, onsite]
        experienceLevel:
          type: ["string", "null"]
        salaryMin:
          type: ["integer", "null"]
        salaryMax:
          type: ["integer", "null"]
        salaryCurrency:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        skills:
          type: array
          items:
            type: string
        source:
          type: string
        applyUrl:
          type: ["string", "null"]
        isNew:
          type: boolean
        isHot:
          type: boolean
        matchScore:
          type: ["integer", "null"]
        postedAt:
          type: ["string", "null"]
        expiresAt:
          type: ["string", "null"]
        createdAt:
          type: string
        isSaved:
          type: boolean
    JobListResponse:
      type: object
      required: [jobs, total, page, limit]
      properties:
        jobs:
          type: array
          items:
            $ref: "#/components/schemas/Job"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...
    SavedJobResult:
      type: object
      properties:
        saved:
          type: boolean
        jobId:
          type: integer

    # ─── APPLICATION ─────────────────────────────────────────────────────────
...
    Resume:
      type: object
      required: [id, userId, name, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: string
        name:
          type: string
        fileName:
          type: ["string", "null"]
        fileUrl:
          type: ["string", "null"]
        fileType:
          type: ["string", "null"]
        version:
          type: integer
        isDefault:
          type: boolean
        content:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    ResumeInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        fileName:
          type: ["string", "null"]
        fileUrl:
          type: ["string", "null"]
        fileType:
          type: ["string", "null"]
        content:
          type: ["string", "null"]
        isDefault:
          type: boolean
    ResumeUpdate:
      type: object
      properties:
        name:
          type: ["string", "null"]
        isDefault:
          type: ["boolean", "null"]
        content:
          type: ["string", "null"]

    # ─── ATS ─────────────────────────────────────────────────────────────────
...
    Notification:
      type: object
      required: [id, userId, type, title, isRead, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: string
        type:
          type: string
          enum: [new_job_match, remote_job, company_posting, application_update, ats_complete]
        title:
          type: string
        message:
          type: ["string", "null"]
        jobId:
          type: ["integer", "null"]
        isRead:
          type: boolean
        createdAt:
          type: string

    # ─── ANALYTICS ───────────────────────────────────────────────────────────
...